// src/app/demo/components/pages/enum/degree.enum.ts
export enum Degree {
    Honours_Agriculture_Plantation_Engineering = "Bachelor of Technology (BTech) Honours in Agriculture and Plantation Engineering",
    Industrial_Studies_Agriculture = "Bachelor of Industrial Studies Honours - Agriculture",
    Software_Engineering = "Bachelor of Software Engineering Honours",
    Electronic_Communication_Engineering = "Bachelor of Technology - Electronic and Communication Engineering",
    Mechanical_Engineering = "Bachelor of Technology - Mechanical Engineering",
    Mechatronics_Engineering = "Bachelor of Technology - Mechatronics Engineering",
    Electrical_Engineering = "Bachelor of Technology - Electrical Engineering",
    Computer_Engineering = "Bachelor of Technology - Computer Engineering",
    Civil_Engineering = "Bachelor of Technology - Civil Engineering",
    Honours_Engineering_Textile_Clothing = "Bachelor of Technology Honours in Engineering – Textile & Clothing",
    Industrial_Studies_Textile_Manufacture_Specialization = "Bachelor of Industrial Studies Honours – Textile Manufacture Specialization",
    Industrial_Studies_Fashion_Design_Product_Development = "Bachelor of Industrial Studies Honours – Fashion Design and Product Development",
    Industrial_Studies_Apparel_Production_Management = "Bachelor of Industrial Studies Honours – Apparel Production and Management"
}